feat: create new button interaction handler system
Add ButtonManager abstract class to oversee all button interactions from commands.
The new system with the decoupled ButtonInteractionHandler permits buttons to be reusable across bot's uptimes.
Move Verifier button interaction logic to a new subclass of ButtonManger for the new system.
