+++ src/base/classes/Handler.ts
@@ -6,6 +6,7 @@ import type Command from "./Command.js";
 import type SubCommand from "./SubCommand.js";
 import { createRequire } from "module";
 import { fileURLToPath } from "url";
+import type ButtonManager from "./ButtonManager.js";
 const require = createRequire(import.meta.url);
 
 export default class Handler {
@@ -51,6 +52,7 @@ export default class Handler {
         });
     }
 
+
     async loadCommands() {
         const dirname = path.dirname(fileURLToPath(import.meta.url));
         let foldersPath: string = "/";
@@ -81,9 +83,38 @@ export default class Handler {
                 this.client.commands.set(command.name, command as Command);
             }
             
-            return delete require.cache[require.resolve(file)]
+            return delete require.cache[require.resolve(file)];
         });
     }
+
+
+    async loadButtonManagers() {
+        const dirname = path.dirname(fileURLToPath(import.meta.url));
+        let foldersPath: string = "/";
+        for (let segment of dirname.split(path.sep)) {
+            if (segment == "base") {
+                foldersPath = path.join(foldersPath, "interactionMangers", "buttonManagers");
+                break;
+            }
+            foldersPath = path.join(foldersPath, segment);
+        }
+
+
+        // files is a list of absolute file paths
+        const files: string[] = this.getFiles(foldersPath);
+        files.map(async (file: string) => {
+            //construct new instance of Command | SubCommand of the file
+            const buttonManager: ButtonManager = new (await import(file)).default(this.client);
+            
+            if (!buttonManager.name) {
+                return delete require.cache[require.resolve(file)] && console.log(`${file.split("/").pop()} does not have a name.`);
+            }
+
+            this.client.buttonManagers.set(buttonManager.name, buttonManager);
+            return delete require.cache[require.resolve(file)];
+        });
+    }
+
 
 
     //baseDirName must the name of a sub dir under src
