+++ src/base/interfaces/IButtonManagerOptions.ts
@@ -1,0 +1,3 @@
+export default interface IButtonManagerOptions {
+    name: string;
+}
